Tested and Fixed MinionSquad, Added TriggerSensor2
I tested MinionSquad to make sure what we have so far works. There was an error in the Minion class when ChangeSquad was called from Start(), but I fixed it.

I also removed SquadManager in favor of just having that code inside of the SquadFormation class.
