+++ InsidiaGame/Assets/Scripts/TriggerSensor2.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: d548d55ea16c5d742b6fb48a5c59f1d4
+timeCreated: 1507687613
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
