+++ InsidiaGame/Assets/Scripts/Util/ISensorListener.cs
@@ -1,0 +1,10 @@
+﻿using UnityEngine;
+using System;
+using System.Collections;
+
+public interface ISensorListener
+{
+    void OnSensorEnter(GameObject other);
+
+    void OnSensorExit(GameObject other);
+}
