+++ InsidiaGame/Assets/Scripts/Util/ISensorListener.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: a87a857cb60e0954c9a3adf24b151702
-timeCreated: 1507406641
+guid: 160ff9781b9e5334394b8ca44c3612eb
+timeCreated: 1507687613
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
